% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_td.R
\name{download_jar}
\alias{download_jar}
\title{Download td-spark jar}
\usage{
download_jar(dest_path = NULL)
}
\arguments{
\item{dest_path}{The destination path where jar will be downloaded to.}
}
\description{
Since file size of td-spark jar is larger than GitHub maximum size,
this command enables to download jar within sparklytd directory.
}
