% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_td.R
\name{spark_read_td}
\alias{spark_read_td}
\title{Read a Treasure Data table into a Spark DataFrame}
\usage{
spark_read_td(sc, name, source, options = list(), repartition = 0,
  memory = TRUE, overwrite = TRUE)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{The name to assign to the newly generated table.}

\item{source}{Source name of the table on TD. Example: \samp{"sample_datasets.www_access"}}

\item{options}{A list of strings with additional options.}

\item{repartition}{The number of partitions used to distribute the
generated table. Use 0 (the default) to avoid partitioning.}

\item{memory}{Boolean; should the data be loaded eagerly into memory? (That
is, should the table be cached?)}

\item{overwrite}{Boolean; overwrite the table with the given name if it
already exists?}
}
\description{
Read a Treasure Data table into a Spark DataFrame
}
\details{
You can read TD table through td-spark. You have to set \code{spark.td.apikey},
\code{spark.serializer} appropreately.
}
\seealso{
Other Spark serialization routines: \code{\link{spark_write_td}}
}
\concept{Spark serialization routines}
